# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Campus Planner Backend",
    "description": "Backend API for AI Campus Planner",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_projects__project_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project_api_projects__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/requirements": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Requirements",
        "operationId": "create_requirements_api_projects__project_id__requirements_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementsSaveRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Requirements",
        "operationId": "get_requirements_api_projects__project_id__requirements_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/buildings": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Building",
        "operationId": "create_building_api_projects__project_id__buildings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildingCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Buildings",
        "operationId": "get_buildings_api_projects__project_id__buildings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildingListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/constraints": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Constraint",
        "operationId": "create_constraint_api_projects__project_id__constraints_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstraintResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Constraints",
        "operationId": "get_constraints_api_projects__project_id__constraints_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstraintListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildingCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "depth": {
            "type": "number",
            "title": "Depth"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "floorCount": {
            "type": "integer",
            "title": "Floorcount"
          },
          "requiredCount": {
            "type": "integer",
            "title": "Requiredcount",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "type",
          "zone",
          "width",
          "depth",
          "height",
          "floorCount"
        ],
        "title": "BuildingCreateRequest"
      },
      "BuildingListResponse": {
        "properties": {
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/BuildingResponse"
            },
            "type": "array",
            "title": "Buildings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "buildings"
        ],
        "title": "BuildingListResponse"
      },
      "BuildingResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "projectId": {
            "type": "integer",
            "title": "Projectid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "depth": {
            "type": "number",
            "title": "Depth"
          },
          "height": {
            "type": "number",
            "title": "Height"
          },
          "floorCount": {
            "type": "integer",
            "title": "Floorcount"
          },
          "requiredCount": {
            "type": "integer",
            "title": "Requiredcount",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "projectId",
          "name",
          "type",
          "zone",
          "width",
          "depth",
          "height",
          "floorCount"
        ],
        "title": "BuildingResponse"
      },
...
      "ConstraintCreateRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "sourceId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          },
          "targetId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetid"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "ConstraintCreateRequest"
      },
      "ConstraintListResponse": {
        "properties": {
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/ConstraintResponse"
            },
            "type": "array",
            "title": "Constraints"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "constraints"
        ],
        "title": "ConstraintListResponse"
      },
      "ConstraintResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "sourceId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          },
          "targetId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetid"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "ConstraintResponse"
      },
      "EntranceSchema": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "width": {
            "type": "number",
            "title": "Width"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "width"
        ],
        "title": "EntranceSchema"
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreateRequest"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "DRAFT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectResponse"
      },
      "ProjectUpdateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectUpdateRequest"
      },
      "RequirementsResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "projectId": {
            "type": "integer",
            "title": "Projectid"
          },
          "siteWidth": {
            "type": "number",
            "title": "Sitewidth"
          },
          "siteHeight": {
            "type": "number",
            "title": "Siteheight"
          },
          "minGreenPercent": {
            "type": "number",
            "title": "Mingreenpercent"
          },
          "minParkingPercent": {
            "type": "number",
            "title": "Minparkingpercent"
          },
          "minRoadWidth": {
            "type": "number",
            "title": "Minroadwidth"
          },
          "minBuildingGap": {
            "type": "number",
            "title": "Minbuildinggap"
          },
          "entrances": {
            "items": {
              "$ref": "#/components/schemas/EntranceSchema"
            },
            "type": "array",
            "title": "Entrances"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "projectId",
          "siteWidth",
          "siteHeight",
          "minGreenPercent",
          "minParkingPercent",
          "minRoadWidth",
          "minBuildingGap"
        ],
        "title": "RequirementsResponse"
      },
      "RequirementsSaveRequest": {
        "properties": {
          "siteWidth": {
            "type": "number",
            "title": "Sitewidth"
          },
          "siteHeight": {
            "type": "number",
            "title": "Siteheight"
          },
          "minGreenPercent": {
            "type": "number",
            "title": "Mingreenpercent"
          },
          "minParkingPercent": {
            "type": "number",
            "title": "Minparkingpercent"
          },
          "minRoadWidth": {
            "type": "number",
            "title": "Minroadwidth"
          },
          "minBuildingGap": {
            "type": "number",
            "title": "Minbuildinggap"
          },
          "entrances": {
            "items": {
              "$ref": "#/components/schemas/EntranceSchema"
            },
            "type": "array",
            "title": "Entrances"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "siteWidth",
          "siteHeight",
          "minGreenPercent",
          "minParkingPercent",
          "minRoadWidth",
          "minBuildingGap"
        ],
        "title": "RequirementsSaveRequest"
      },
...
      "ValidationErrorResponse": {
        "title": "ValidationErrorResponse",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string",
            "example": "VALIDATION_ERROR"
          }
        },
        "required": [
          "detail"
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}